% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp.R
\name{mtp_population}
\alias{mtp_population}
\title{Generate MTP population from combined}
\usage{
mtp_population(combined)
}
\arguments{
\item{combined}{Output from combined fitting task
rtm_inference_pmcmc_spim_fits2_combined}
}
\description{
Generate MTP population from combined
}
