% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{plot_rt_dist}
\alias{plot_rt_dist}
\title{Plot Rt distribution}
\usage{
plot_rt_dist(
  npi_key,
  xlim,
  ylim,
  labels = NULL,
  legend_ncol = 1,
  npi_key2 = NULL
)
}
\arguments{
\item{npi_key}{data.frame type object with columns:
nations, npi (scenario), Rt (mean), Rt_sd}

\item{xlim}{Passed to \code{plot}}

\item{ylim}{Passed to \code{plot}}

\item{labels}{Passed to \link{legend} \code{legend} parameter}

\item{legend_ncol}{Passed to \link{legend} \code{ncol} parameter}

\item{npi_key2}{Optional second npi_key that will be plotted with dashed
lines}
}
\description{
Plot Rt distribution over time for various scenarios
}
