% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\name{spim_add_forecast}
\alias{spim_add_forecast}
\title{Add forecasts to combined fit object}
\usage{
spim_add_forecast(dat, forecast_days)
}
\arguments{
\item{dat}{A combined fit object}

\item{forecast_days}{Number of days of forecasts to add}
}
\value{
A combined fit object with forecasts added in
}
\description{
Add forecasts to combined fit object
}
