% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{spim_add_par}
\alias{spim_add_par}
\alias{spim_add_par_beta}
\title{Add parameter}
\usage{
spim_add_par(
  pars,
  name,
  initial,
  min,
  max,
  proposal_variance,
  prior,
  discrete = FALSE,
  include = TRUE
)

spim_add_par_beta(pars)
}
\arguments{
\item{pars}{Original \code{spim_pars_pmcmc} object}

\item{name}{Name of the new parameter}

\item{initial}{Initial value for the new parameter}

\item{min}{Minimum value}

\item{max}{Maximum value}

\item{proposal_variance}{Variance for the proposal VCV}

\item{prior}{List of values for the prior (must include "type",
then other values for the prior data.frame)}

\item{discrete}{Logical, indicates if discrete}

\item{include}{Logical, indicates if fitted}
}
\description{
Add a parameter to a \code{spim_pars_mcmc} object
}
