% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_simulation.R
\name{spim_beta_mult_schools}
\alias{spim_beta_mult_schools}
\title{Calculate multiplicative beta for schools}
\usage{
spim_beta_mult_schools(dates, schedule, modifier, region)
}
\arguments{
\item{dates}{A sircovid date, fractional dates (e.g. 100.25) are
allowed.}

\item{schedule}{A data.frame of school open/closed
information. Required columns are \code{region}, \code{date} and
\code{schools}; really this should have been sorted out by
\link{spim_simulate_control}}

\item{modifier}{The amount to \emph{decrease} beta by when schools are
closed for holidays etc. A value of 0.15 is a 15\% reduction in
contacts, or a multiplicative beta of 0.85}

\item{region}{The region to subset from the schedule.}
}
\value{
A vector the same length as \code{dates} with the scaling
factor.
}
\description{
Calculate multiplicative beta scaling factor for schools.  We
assume here that schools has come from
\link{spim_simulate_control} but we may add validation later.
}
