% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_simulation.R
\name{spim_beta_mult_seasonality}
\alias{spim_beta_mult_seasonality}
\title{Calculate multiplicative beta due to seasonality}
\usage{
spim_beta_mult_seasonality(dates, date_peak, value)
}
\arguments{
\item{dates}{A sircovid date, fractional dates (e.g. 100.25) are
allowed.}

\item{date_peak}{Date of peak as a
\link[sircovid:sircovid_date]{sircovid::sircovid_date} value (a single number)}

\item{value}{Size of the peak to annual average difference. A
value of 0.1 will produce multiplicative values that range from
0.9 to 1.1.}
}
\value{
A vector the same length as \code{dates} with the scaling
factor.
}
\description{
Calculate multiplicative beta scaling factor due to seasonality.
}
