% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_calc_seasonality}
\alias{spim_calc_seasonality}
\title{Seasonality}
\usage{
spim_calc_seasonality(date, seasonality_date_peak, seasonality)
}
\arguments{
\item{date}{date in sircovid date format}

\item{seasonality_date_peak}{date of peak in sircovid date format}

\item{seasonality}{peak to annual average difference}
}
\value{
a value or vector of values with seasonal effect for each date
}
\description{
Seasonality
}
