% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{spim_check_model_type}
\alias{spim_check_model_type}
\title{Check a model type}
\usage{
spim_check_model_type(model_type)
}
\arguments{
\item{model_type}{String}
}
\description{
Check that a model type is valid (i.e., one of \code{NB} or \code{BB})
}
