% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{spim_check_region}
\alias{spim_check_region}
\title{Validate and expand regions}
\usage{
spim_check_region(region, multiregion)
}
\arguments{
\item{region}{A single string indicating a region or group of
regions.  For \code{multiregion = FALSE} this must be a suitable NHS
region.  When \code{multiregion = TRUE} this must be one of \code{all},
\code{england} or \code{test}}

\item{multiregion}{Logical, indicating if this is a multiregion fit}
}
\value{
A character vector of regions
}
\description{
Validate (and possibly expand) a region for a fit.
}
