% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{spim_check_sircovid_model}
\alias{spim_check_sircovid_model}
\title{Check a model type}
\usage{
spim_check_sircovid_model(sircovid_model)
}
\arguments{
\item{sircovid_model}{String}
}
\description{
Check that a model type is valid (i.e., one of \code{NB} or \code{BB})
}
