% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\name{spim_combined_load_multiregion}
\alias{spim_combined_load_multiregion}
\title{Load multiregion fits as a combined fit object}
\usage{
spim_combined_load_multiregion(path)
}
\arguments{
\item{path}{Directory name. Within here we expect to see
\code{fits.rds}}
}
\value{
A combined fit object
}
\description{
Load multiregion fits as a combined fit object
}
