% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{spim_control_cores}
\alias{spim_control_cores}
\title{Return the number of cores}
\usage{
spim_control_cores()
}
\value{
An integer
}
\description{
Return the number of available cores, looking at \code{CONTEXT_CORES},
\code{MC_CORES} and \code{mc.cores} in turn
}
