% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_create_rt_scenario}
\alias{spim_create_rt_scenario}
\title{Create Rt schedule scenario}
\usage{
spim_create_rt_scenario(sched, region, scenario, schedule)
}
\arguments{
\item{sched}{A data frame with school closure schedule. Expect
columns: nation; year; month; day; schools ('open' if open on that day
otherwise 'closed')}

\item{region}{Region to filter csv by}

\item{scenario}{Name of scenario to create}

\item{schedule}{data.frame with columns: date (date of npi change),
npi (name of corresponding npi in npi_key), nation}
}
\description{
Create a scenario for Rt schedule
}
