% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_expand_grid}
\alias{spim_expand_grid}
\title{Create expanded run grid}
\usage{
spim_expand_grid(..., full_run = FALSE, prefix = "Grid_")
}
\arguments{
\item{...}{named variables to expand over, should be in
\code{simulation_central_analysis}, omitted variables will take central value}

\item{full_run}{If \code{TRUE} saves trajectories for expanded scenarios, this
should very rarely be \code{TRUE}, change default with care as this will lead
to massive objects}

\item{prefix}{prefix for analysis name, prefixes row number}
}
\value{
A grid of scenarios to run
}
\description{
Create expanded run grid for simulation
}
