% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_spim.R
\name{spim_extract_admissions_by_age}
\alias{spim_extract_admissions_by_age}
\title{Extract admissions by age}
\usage{
spim_extract_admissions_by_age(dat)
}
\arguments{
\item{dat}{Combined data set}
}
\description{
Extract admissions by age
}
