% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_spim.R
\name{spim_extract_alos}
\alias{spim_extract_alos}
\title{Extract average length of hospital stay}
\usage{
spim_extract_alos(dat, regions, date)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{String or vector of strings with region names for output}

\item{date}{Latest data point date}
}
\description{
Extract current ALOS by region
}
