% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_spim.R
\name{spim_extract_variants_rt}
\alias{spim_extract_variants_rt}
\title{Extract multivariant Rt}
\usage{
spim_extract_variants_rt(dat, type_rt)
}
\arguments{
\item{dat}{Combined data set}

\item{type_rt}{Either 'Rt_general' or 'eff_Rt_general'}
}
\description{
Extract current Rt by variant
}
