% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_process.R
\name{spim_fit_process}
\alias{spim_fit_process}
\title{Process a fit}
\usage{
spim_fit_process(samples, control, admissions, rtm, parameters)
}
\arguments{
\item{samples}{The pmcmc_samples object from \link[mcstate:pmcmc]{mcstate::pmcmc}}

\item{control}{The forecast control from \link{spim_control}}

\item{admissions}{The admissions data set from
\link{spim_data_admissions}}

\item{rtm}{The rtm data set}

\item{parameters}{The parameter specification
(\link{spim_pars_pmcmc_load})}
}
\description{
Process fit data
}
