% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{spim_fit_run}
\alias{spim_fit_run}
\title{Run fit}
\usage{
spim_fit_run(pars, filter, control)
}
\arguments{
\item{pars}{Parameters}

\item{filter}{A particle filter object, typically from
\link{spim_particle_filter}}

\item{control}{A \link[mcstate:pmcmc_control]{mcstate::pmcmc_control} object, typically the
\code{pmcmc} element of the result of \code{\link[=spim_control]{spim_control()}}}
}
\value{
A \code{mcstate_pmcmc} object
}
\description{
Run the fit (using pmcmc); this may take a very long time
}
