% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_process.R
\name{spim_lancelot_fit_process}
\alias{spim_lancelot_fit_process}
\title{Process a fit for \code{lancelot} model}
\usage{
spim_lancelot_fit_process(
  samples,
  parameters,
  data,
  control,
  random_sample = TRUE
)
}
\arguments{
\item{samples}{The pmcmc_samples object from \link[mcstate:pmcmc]{mcstate::pmcmc}}

\item{parameters}{The parameter specification
(\link{spim_pars_pmcmc_load})}

\item{data}{Data sets used in fitting, via
\link{spim_fit_process_data}}

\item{control}{The forecast control from \link{spim_control}}

\item{random_sample}{Logical parameter, if \code{TRUE} will obtain the
posterior samples via random sampling, otherwise thinning will
be used}
}
\description{
Process fit data for \code{lancelot} model
}
