% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp_new.R
\name{spim_mtp_age_vaccine_outputs}
\alias{spim_mtp_age_vaccine_outputs}
\title{MTP simulation outputs by age and vaccination class}
\usage{
spim_mtp_age_vaccine_outputs(res, region = "england")
}
\arguments{
\item{region}{A string, region for which outputs will be plotted}

\item{dat}{Output from MTP simulation}
}
\description{
Prepare outputs by age and vaccination class for plotting
}
