% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp_new.R
\name{spim_mtp_summary_to_template}
\alias{spim_mtp_summary_to_template}
\title{Port MTP outcomes to template}
\usage{
spim_mtp_summary_to_template(
  summary_tidy,
  date,
  run_grid,
  combined,
  spim_state_names
)
}
\arguments{
\item{summary_tidy}{A tibble containing MTP projections}

\item{date}{The date to start output from (not necessarily the model date)}

\item{run_grid}{The grid of scenarios ran}

\item{combined}{Output from combined fitting task
rtm_inference_pmcmc_spim_fits2_combined}

\item{spim_state_names}{Named vector of states to be extracted for SPI-M}
}
\description{
Port MTP outcomes to template
}
