% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multivariant_rt.R
\name{spim_multivariant_rt_plot}
\alias{spim_multivariant_rt_plot}
\title{Create multivariant Rt plot}
\usage{
spim_multivariant_rt_plot(
  dat,
  date,
  date_restart,
  last_beta_days_ago = 21,
  region = "england",
  rt_type = "eff_Rt_general"
)
}
\arguments{
\item{dat}{Main fitting outputs object}

\item{date}{Date of the analysis}

\item{date_restart}{Date for multivariant model restart}

\item{last_beta_days_ago}{Integer for setting last beta changepoint in
relation to date of analysis}

\item{region}{A string for the name of the region to plot}

\item{rt_type}{A string, must be one of eff_Rt_general or Rt_general}
}
\value{
A ggplot2 object for multivariant Rt
}
\description{
Create multivariant Rt plot
}
