% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{spim_pars}
\alias{spim_pars}
\title{Create parameters object}
\usage{
spim_pars(
  date,
  region,
  model_type,
  multistrain,
  beta_date,
  vaccination,
  parameters
)
}
\arguments{
\item{date}{The date of the data. This will be the final date}

\item{region}{The region that we are fitting to. This will either
be one of the 10 NHS regions (one of
\code{sircovid::regions("all")}), or an argument to that function
(e.g., \code{all})}

\item{model_type}{Either "BB" (beta-binomial) or "NB" (negative
binomial). We use and exclude different data types depending on
this argument.}

\item{multistrain}{Logical, indicating if the model is a
"multistrain" model allowing for mulitiple competing strains.}

\item{beta_date}{A vector of date (strings) for the beta
parameters. Must align with parameters}

\item{vaccination}{Vacination data, from
\link{spim_vaccination_data}}

\item{parameters}{Parameter information, from
\link{spim_pars_pmcmc_load}}
}
\value{
An \link[mcstate:pmcmc_parameters]{mcstate::pmcmc_parameters} object which can be used
with \link[mcstate:pmcmc]{mcstate::pmcmc}
}
\description{
Create parameters object for use with \link[mcstate:pmcmc]{mcstate::pmcmc}. This
function is very data-hungry as it (alongside
\code{\link[=spim_data]{spim_data()}}) is the main point at which data enters
into the model. Here we end up setting a lot of data that are are
not fitting to, but taking as fixed inputs (vaccination over time)
as well as information on the parameters that the model will vary
(their names, ranges, priors and their proposal distributions).
}
