% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{spim_pars_beta}
\alias{spim_pars_beta}
\title{Create vector of beta dates}
\usage{
spim_pars_beta(date)
}
\arguments{
\item{date}{Todays dates (last date is set two weeks ago)}
}
\description{
Create vector of beta dates
}
