% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{spim_pars_check_beta_date}
\alias{spim_pars_check_beta_date}
\title{Validate vector of beta dates}
\usage{
spim_pars_check_beta_date(beta_date)
}
\arguments{
\item{beta_date}{Vector of beta dates}
}
\description{
Validate vector of beta dates
}
