% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{spim_pars_pmcmc_save}
\alias{spim_pars_pmcmc_save}
\title{Write out parameters}
\usage{
spim_pars_pmcmc_save(p, path)
}
\arguments{
\item{p}{A list with elements \code{info}, \code{prior} and \code{proposal}}

\item{path}{Directory to write to}
}
\value{
Nothing, called for side effects
}
\description{
Write pmcmc paramters out after fitting and aggregation (inverse
of \link{spim_pars_pmcmc_load})
}
