% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_admissions_by_age}
\alias{spim_plot_admissions_by_age}
\title{Plot admissions by age}
\usage{
spim_plot_admissions_by_age(dat, region)
}
\arguments{
\item{dat}{Combined data set}

\item{region}{Region to plot}
}
\description{
Plot admissions by age
}
