% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_alos}
\alias{spim_plot_alos}
\title{Plot average length of stay over time}
\usage{
spim_plot_alos(dat, regions, ymin = NULL, ymax = NULL, forecast_until = NULL)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{ymax}{Maximum percentage on y-axis}

\item{forecast_until}{Optional date to forecast till}
}
\description{
Plot average length of stay over time
}
