% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_plot_check_doses}
\alias{spim_plot_check_doses}
\title{Check doses given out from simulation}
\usage{
spim_plot_check_doses(doses)
}
\arguments{
\item{doses}{Output from \link{spim_calculate_doses}}
}
\description{
Diagnostic plots for checking doses given out from simulation over
all analyses and age groups
}
