% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_plot_check_rt}
\alias{spim_plot_check_rt}
\title{Check Rt from simulation}
\usage{
spim_plot_check_rt(
  summary_state,
  dates,
  combined_state = NULL,
  Rt_state = "eff_Rt_general_both"
)
}
\arguments{
\item{summary_state}{State from simulation summary object}

\item{dates}{Dates to plot vertical lines at}

\item{combined_state}{Optional state from combined object}

\item{Rt_state}{Rt states to plot, default is effective average Rt}
}
\description{
Diagnostic plots for checking Rt from simulation
}
