% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_plot_check_state_by_age}
\alias{spim_plot_check_state_by_age}
\title{Check states by age from simulation}
\usage{
spim_plot_check_state_by_age(summary_agestate, ana, scen)
}
\arguments{
\item{summary_agestate}{State by age from simulation summary object}

\item{ana}{Analysis to check, usually central}

\item{scen}{Scenario to check, usually central}
}
\description{
Diagnostic plots for checking states by age from simulation
}
