% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_plot_check_total_doses}
\alias{spim_plot_check_total_doses}
\title{Check total doses given out from simulation}
\usage{
spim_plot_check_total_doses(doses)
}
\arguments{
\item{doses}{Output from \link{spim_calculate_doses}}
}
\description{
Diagnostic plots for checking total doses given out in simulation
}
