% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_plot_check_uptake}
\alias{spim_plot_check_uptake}
\title{Check dose uptake from simulation}
\usage{
spim_plot_check_uptake(doses)
}
\arguments{
\item{doses}{Output from \link{spim_calculate_doses}}
}
\description{
Diagnostic plots for checking dose uptake from simulation
}
