% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_cumulative_attack_rate}
\alias{spim_plot_cumulative_attack_rate}
\title{Plot cumulative attack rate}
\usage{
spim_plot_cumulative_attack_rate(dat, regions)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}
}
\description{
Plot cumulative attack rate
}
