% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_plot_daily_infections}
\alias{spim_plot_daily_infections}
\title{Check daily infections from simulation}
\usage{
spim_plot_daily_infections(summary_state, ana, scen)
}
\arguments{
\item{summary_state}{State from simulation summary object}

\item{ana}{Analysis to plot}

\item{scen}{Scenario to plot}
}
\description{
Diagnostic plots for checking daily infections and proportion of strain 2
}
