% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_effective_susceptible}
\alias{spim_plot_effective_susceptible}
\title{Plot effective susceptibles}
\usage{
spim_plot_effective_susceptible(dat, regions, strain_names, strain_dates)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{strain_names}{Names of strains}

\item{strain_dates}{Dates of strains introduced into model}
}
\description{
Plot effective susceptibles
}
