% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{spim_plot_fit_forecasts}
\alias{spim_plot_fit_forecasts}
\alias{spim_plot_fit_traces}
\alias{spim_plot_fit_posteriors}
\title{Plots for fit}
\usage{
spim_plot_fit_forecasts(samples, region = NULL)

spim_plot_fit_traces(samples)

spim_plot_fit_posteriors(samples, region = NULL)
}
\arguments{
\item{samples}{The samples}

\item{region}{The region. Leave this \code{NULL} except for
multi-region fits}
}
\value{
Nothing, called for side effect
}
\description{
Plots for fits
}
