% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined_forest.R
\name{spim_plot_forest}
\alias{spim_plot_forest}
\title{Forest plot}
\usage{
spim_plot_forest(dat, regions = NULL, plot_type = "all")
}
\arguments{
\item{dat}{Combined data from \link{spim_combined_load}}

\item{regions}{Vector of regions to plot. By default, all regions
found in \code{dat} will be used (except aggregate regions, such as
england/uk)}

\item{plot_type}{The type of parameters to plot: \code{all} would plot
all parameters, \code{betas} would just plot betas, \code{non_betas} would
plot all non-beta parameters}
}
\value{
Nothing, called for side effects
}
\description{
Forest plot
}
