% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_incidence}
\alias{spim_plot_incidence}
\title{Plot incidence}
\usage{
spim_plot_incidence(dat, regions, per_1000 = FALSE)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{per_100}{Logical, indicating if we plot incidence per 1000 people}
}
\description{
Plot incidence
}
