% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_infection_status}
\alias{spim_plot_infection_status}
\title{Plot infection status}
\usage{
spim_plot_infection_status(dat, regions)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}
}
\description{
Plot infection status
}
