% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_log_traj_by_age}
\alias{spim_plot_log_traj_by_age}
\title{Plot log trajectories by age}
\usage{
spim_plot_log_traj_by_age(dat, regions, yield, model_type)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{yield}{Which kind of age-specific trajectories to add, can be
"pillar2", "admissions" or "deaths"}

\item{model_type}{Which model is being run, either "BB" or "NB"}
}
\description{
Plot log trajectories by age
}
