% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_pillar2_cases}
\alias{spim_plot_pillar2_cases}
\title{Plot Pillar 2 cases}
\usage{
spim_plot_pillar2_cases(
  dat,
  regions,
  date_min,
  age_band = NULL,
  add_betas = FALSE
)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{date_min}{Starting date for plot}

\item{add_betas}{Logical, indicating if we should add betas}
}
\description{
Plot Pillar 2 cases
}
