% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_prop_susceptible}
\alias{spim_plot_prop_susceptible}
\title{Plot proportion susceptible}
\usage{
spim_plot_prop_susceptible(dat, regions, ymin)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{forecast_until}{Optional date to forecast till}
}
\description{
Plot proportion susceptible
}
