% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_Rt}
\alias{spim_plot_Rt}
\title{Plot Rt}
\usage{
spim_plot_Rt(dat, rt_type, forecast_until = NULL)
}
\arguments{
\item{dat}{Combined data set}

\item{rt_type}{One of the valid Rt types (e.g., eff_Rt_all)}

\item{forecast_until}{Optional date to forecast till}
}
\description{
Plot Rt
}
