% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_plot_seasonality}
\alias{spim_plot_seasonality}
\title{Plot sesonality over time}
\usage{
spim_plot_seasonality(peak_date = as.Date("2020-02-15"), seasonality = 0.1)
}
\arguments{
\item{peak_date}{Date where seasonal multiplier is highest}

\item{seasonality}{Seasonal multiplier}
}
\description{
Plot seasonality trends over time
}
