% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multivariant.R
\name{spim_plot_seeding_date}
\alias{spim_plot_seeding_date}
\title{Create variant seeding date plot}
\usage{
spim_plot_seeding_date(dat)
}
\arguments{
\item{dat}{Main fitting outputs object}
}
\value{
A ggplot2 object for fitted seeding date by region
}
\description{
Create variant seeding date plot
}
