% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_serology}
\alias{spim_plot_serology}
\title{Plot serology}
\usage{
spim_plot_serology(dat, sero_flow, ymax)
}
\arguments{
\item{dat}{Combined data set}

\item{sero_flow}{Number identifying which serology flow to use (1 or 2)}

\item{ymax}{Maximum percentage on y-axis}
}
\description{
Plot serology
}
