% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_trajectories_by_age}
\alias{spim_plot_trajectories_by_age}
\title{Plot trajectories by age}
\usage{
spim_plot_trajectories_by_age(
  dat,
  regions,
  what,
  date_min = NULL,
  age_band = NULL,
  with_forecast = TRUE,
  add_betas = FALSE
)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{what}{Trajectory to plot}

\item{date_min}{Starting date for plot}

\item{age_band}{Age band to plot}

\item{with_forecast}{Logical, indicating if we should add the forecast}

\item{add_betas}{Logical, indicating if we should add betas}
}
\description{
Plot trajectories by age
}
