% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_unvaccinated_susceptible}
\alias{spim_plot_unvaccinated_susceptible}
\title{Plot susceptible}
\usage{
spim_plot_unvaccinated_susceptible(dat, regions, date)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{date}{Current date for extracting susceptible proportion}
}
\description{
Plot susceptible unvaccinated proportion
}
