% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multivariant.R
\name{spim_plot_vaccine_figures}
\alias{spim_plot_vaccine_figures}
\title{Create plot for vaccine paper figure 1}
\usage{
spim_plot_vaccine_figures(dat, date, date_restart, manuscript = TRUE)
}
\arguments{
\item{dat}{Main fitting outputs object}

\item{date}{Date of the analysis}

\item{date_restart}{Date for multivariant model restart}

\item{manuscript}{Logical indicating if plotting for manuscript}
}
\value{
A ggplot2 object for fit to VOC proportion data
}
\description{
Create plot for vaccine paper figure 1
}
