% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_vaccine_status}
\alias{spim_plot_vaccine_status}
\title{Plot vaccine status}
\usage{
spim_plot_vaccine_status(dat, regions, vacc_names)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{vacc_names}{Character vector of names of vaccine statuses}
}
\description{
Plot vaccine status
}
