% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_variant}
\alias{spim_plot_variant}
\title{Plot variant}
\usage{
spim_plot_variant(dat, regions, date_min = NULL, date_max = NULL)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{date_min}{Starting date for plot}

\item{date_max}{End date for plot}
}
\description{
Plot variant
}
