% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multivariant.R
\name{spim_plot_variant_transmission}
\alias{spim_plot_variant_transmission}
\title{Create variant transmission advantage plot}
\usage{
spim_plot_variant_transmission(dat)
}
\arguments{
\item{dat}{Main fitting outputs object}
}
\value{
A ggplot2 object for fitted seeding date by region
}
\description{
Create variant transmission advantage plot
}
