% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_plot_voc_range}
\alias{spim_plot_voc_range}
\title{Plot relative strain transmissibility ranges}
\usage{
spim_plot_voc_range(R1, R1_sd, epsilon_range, epsilon_central)
}
\arguments{
\item{R1}{Lognormal mean for strain 1}

\item{R1_sd}{Lognormal standard deviation for strain 1}

\item{epsilon_range}{Relative range for strain 2}

\item{epsilon_central}{Relative mean for strain 2}
}
\description{
Plot relative VOC mean and sd
}
