% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{spim_pmcmc}
\alias{spim_pmcmc}
\title{Run a pmcmc}
\usage{
spim_pmcmc(pars, filter, control)
}
\arguments{
\item{pars}{Parameters, from \link{spim_pars}}

\item{filter}{A particle filter object, typically from
\link{spim_particle_filter}}

\item{control}{A \link[mcstate:pmcmc_control]{mcstate::pmcmc_control} object, typically the
\code{pmcmc} element of the result of \code{\link[=spim_control]{spim_control()}}}
}
\value{
A \code{mcstate_pmcmc} object
}
\description{
Run a pmcmc; this may take a very long time
}
