% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\name{spim_population}
\alias{spim_population}
\title{Get population from combined}
\usage{
spim_population(combined, ignore_uk = FALSE, by_age = TRUE, group = 1:19)
}
\arguments{
\item{combined}{Combined fits object}

\item{ignore_uk}{If \code{TRUE} population of UK is retured as NA}
}
\value{
data.frame of region/country populations
}
\description{
Get region and country population from a combined fits object
}
