% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_prepare_aggregated_data}
\alias{spim_prepare_aggregated_data}
\title{Prepare aggregated data for plotting}
\usage{
spim_prepare_aggregated_data(path)
}
\arguments{
\item{path}{Path to aggregated rds object containing a named list where
names correspond to regions and each element is a list with \code{full} data
and \code{fitted} data}
}
\value{
Returns a list where elements correspond to regions and \code{fitted}
data is removed. Data processing includes: adding \code{deaths}
column as the sum over all death compartments; fixes when NAs converted to
0s erroneously; and fitted data removed.
}
\description{
Prepare aggregated real data for plotting
}
