% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_prepare_npi_key}
\alias{spim_prepare_npi_key}
\title{Prepare NPI key csv for simulation}
\usage{
spim_prepare_npi_key(path, country)
}
\arguments{
\item{path}{File path to a csv containing columns: nation
(at least one of england, scotland, wales, northern ireland), npi
(name of NPI step for associated schedule), Rt (mean value for Rt),
Rt_sd (standard deviation for Rt)}

\item{country}{If \code{"england"} then all other nations filtered, otherwise
no filtering.}
}
\value{
tibble for passing to \link{spim_prepare_rt_future}
}
\description{
Prepare csv with NPI keys for simulation task
}
