% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_prepare_rt_future}
\alias{spim_prepare_rt_future}
\title{Prepare Rt future for simulation}
\usage{
spim_prepare_rt_future(path, npi_key, start_date, end_date)
}
\arguments{
\item{path}{File path to a csv containing columns: nation
(at least one of england, scotland, wales, northern ireland), scenario
(scenario ID), year (YYYY), month (MM), (DD), npi
(NPI ID to match npi_key$npi)}

\item{start_date}{Start date of simulation, all changes in schedule before
this date are removed.}

\item{end_date}{End date of simulation, all changes in schedule after
this date are removed.}
}
\value{
tibble which is essentially a column binds of npi_key and
Rt_schedule datasets after cleaning
}
\description{
Prepare Rt future grid for simulation task
}
