% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\name{spim_prop_infected}
\alias{spim_prop_infected}
\title{Get proportion infected from combined}
\usage{
spim_prop_infected(
  combined,
  population,
  regions = sircovid::regions("england")
)
}
\arguments{
\item{combined}{Combined fits object}

\item{population}{Population from \link{spim_population}}

\item{regions}{Regions for which to get proportion infected}
}
\value{
data.frame of region/country proportion infected
}
\description{
Get proportion infected from country population from a combined fits object
}
