% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{spim_region_name}
\alias{spim_region_name}
\title{Convert region names}
\usage{
spim_region_name(region, type = "name")
}
\arguments{
\item{region}{Vector of standard region names (london, scotland,
england, uk)}

\item{type}{Convertion type. Current can be one of "name", "code" or
"upper"}
}
\value{
A vector of new names
}
\description{
Convert regions to different types of names
}
