% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_rejuvenatoR}
\alias{spim_rejuvenatoR}
\title{Find daily little r values from big R}
\usage{
spim_rejuvenatoR(
  summary,
  dates,
  scenarios = NULL,
  analyses = NULL,
  wide = FALSE
)
}
\arguments{
\item{summary}{Simulation summary object}

\item{dates}{Dates for which to compute little r}

\item{scenarios}{Scenarios for which to compute little r . If \code{NULL}
computed over all scenarios.}

\item{analyses}{Analyses for which to compute little r . If \code{NULL}
computed over all analyses.}

\item{wide}{If \code{FALSE} (default) returns results in long format as numeric,
otherwise returns wide format with dates as columns and scenarios/analyses
as rows and entries are given as \code{central (low, high)}}
}
\value{
data.frame with daily little r at given dates, scenarios, analyses
}
\description{
Find little r from big R
}
