% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart.R
\name{spim_restart_initial}
\alias{spim_restart_initial}
\title{Create restart initial conditions}
\usage{
spim_restart_initial(state, date, multistrain)
}
\arguments{
\item{state}{Final state at the point of the restart (matrix with
rows corresponding to model state and columns corresponding to
samples)}

\item{date}{The sircovid date of the restart; this will be
converted into the initial step.}

\item{multistrain}{Logical, whether the restart fits are multistrain or not}
}
\description{
Create an mcstate "initial" function from a previous final state
matrix
}
