% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_restart.R
\name{spim_restart_initial_state}
\alias{spim_restart_initial_state}
\title{Load restart data}
\usage{
spim_restart_initial_state(restart, date, round = FALSE)
}
\arguments{
\item{restart}{The loaded restart date (processed with
\link{spim_fit_process} in a previous task). We require that
\code{restart} has a \code{state} element which is list with elements
\code{time} (a sircovid date) and \code{state} (a 3d array of model state
by particle by \code{time}).}

\item{date}{The date that the restart should come from.  Will be
passed through \link[sircovid:sircovid_date]{sircovid::as_sircovid_date} so can be either an
integer sircovid date, an R Date or a string representing an ISO
date.}

\item{round}{Logical parameter, indicates whether or not to round
the main model states. Can be used to ensure that a restart initial
state from a deterministic parent fit can be used in a stochastic
restart fit. Default is FALSE}
}
\value{
A list with the same names as \code{restart}, but with
\code{restart$state} filtered down to a single date's data and
\code{restart$state$restart_date} replaced by the sircovid date
corresponding to the restart.
}
\description{
Load restart state from restart date.  This takes output from a
combined task that has been processed with
\link{spim_fit_process} to return the model state at a given
date.
}
