% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_restart.R
\name{spim_restart_join_parent}
\alias{spim_restart_join_parent}
\title{Join parent and restart fits}
\usage{
spim_restart_join_parent(fit, parent)
}
\arguments{
\item{fit}{The new fit. A list with elements \code{samples}, \code{pmcmc},
etc.}

\item{parent}{The parent information from the restart data created
by \link{spim_fit_process}; this contains elements
\code{trajectories}, \code{rt} and \code{data} and we use these to
create consistent trajectories that cover both the parent and
restart fits.}
}
\value{
A list with the same elements as \code{fit}, but with
concatenated trajectories as appropriate.
}
\description{
Join parent and restart fits together
}
