% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_restart.R
\name{spim_restart_pars}
\alias{spim_restart_pars}
\title{Create restart parameters}
\usage{
spim_restart_pars(pars, pars_parent, restart_date)
}
\arguments{
\item{pars}{The full set of parameters.  This is a list with the
baseline (\code{base}) along with \code{prior}, \code{info} and \code{transform}.}

\item{pars_parent}{The parameters from the parent fit; this must
be a \code{spim_pars_mcmc} object}

\item{restart_date}{The date that restart happens, will be pased
through \code{spimalot::as_sircovid_date}}
}
\value{
An object with the same structure as \code{pars} but with the
\code{prior} element updated and an \code{mcmc} element added.
}
\description{
Create restart parameters
}
