% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{spim_round_to}
\alias{spim_round_to}
\title{Round to given number}
\usage{
spim_round_to(x, y)
}
\arguments{
\item{x}{Number to round}

\item{y}{Number to round to}
}
\description{
Round to nearest given number
}
\examples{
 # Round to nearest 100
 spim_round_to(146, 100)
 # Round to nearest 1000
 spim_round_to(7849, 1000)
}
