% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_run_grid}
\alias{spim_run_grid}
\title{Create scenario run grid}
\usage{
spim_run_grid(
  scenarios,
  csv = NULL,
  expand_grid = NULL,
  force_central = TRUE,
  set_strain_params = TRUE,
  multistrain = TRUE,
  analyses = NULL
)
}
\arguments{
\item{scenarios}{Scenarios to run simulation over}

\item{csv}{Path of csv to load run grid from}

\item{expand_grid}{Optional large grid of scenarios such as from
\link{spim_expand_grid}}

\item{force_central}{If \code{TRUE} (default) then central analysis is always
included as specified in \code{simulation_central_analysis}. This should rarely
be \code{FALSE} as often required for basic checking plots.}

\item{set_strain_params}{If \code{TRUE} automatically sets strain parameters
\code{strain_cross_immunity} and \code{strain_severity_modifier}, which are
currently equivalent to \code{strain_vaccine_efficacy}}

\item{multistrain}{If \code{FALSE} then removes all columns related to a second
strain}

\item{analyses}{If not \code{NULL} then filters grid by given analyses}
}
\value{
A grid of scenarios to run
}
\description{
Create grid of scenarios to run for simulation
}
