% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_scenario_cols}
\alias{spim_scenario_cols}
\title{Return accessible scenario colours}
\usage{
spim_scenario_cols(
  scenarios,
  dark_scenarios = NULL,
  light_scenarios = NULL,
  weight = 0.3,
  palette = "bright",
  highR = TRUE,
  lowR = TRUE,
  preview = FALSE
)
}
\arguments{
\item{scenarios}{Unique scenario names}

\item{dark_scenarios}{Optional unique scenario names that should be same
length as \code{scenarios} as provided and will be the same colours but darker.
Useful if plotting scenarios and their High R counterparts.}

\item{light_scenarios}{Optional unique scenario names that should be same
length as \code{scenarios} as provided and will be the same colours but lighter.
Useful if plotting scenarios and their Low R counterparts.}

\item{weight}{If \code{dark_scenarios} is not \code{NULL} then \code{weight} passed to
\code{mix_cols} to darken colours by mixing with "#000000" (black)}

\item{palette}{Colour palette, passed to \link[khroma:colour]{khroma::colour}}

\item{highR}{If \code{TRUE} then \code{dark_scenarios} is taken to be all scenarios
that contain "High R"; \code{dark_scenarios} should be \code{NULL} if \code{TRUE}}

\item{lowR}{If \code{TRUE} then \code{light_scenarios} is taken to be all scenarios
that contain "Low R"; \code{light_scenarios} should be \code{NULL} if \code{TRUE}}

\item{preview}{If \code{TRUE} then plots the final colour scheme with
\link[khroma:plot]{khroma::plot_scheme}}
}
\description{
Select colours for plotting simulation scenarios
}
