% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_simulate}
\alias{spim_simulate}
\title{Run simulations}
\usage{
spim_simulate(args, combined, rrq = NULL)
}
\arguments{
\item{args}{Arguments returned by \link{spim_simulate_args}}

\item{combined}{Processed combined output returned by
\link{spim_simulate_prepare}}

\item{rrq}{rrq object, if NULL runs locally}
}
\description{
Run simulations locally or with rrq workers
}
