% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_add_all_deaths}
\alias{spim_simulate_add_all_deaths}
\title{Combine all death trajectories}
\usage{
spim_simulate_add_all_deaths(obj, incidence = FALSE)
}
\arguments{
\item{obj}{A simulated object}

\item{incidence}{Logical, whether you are combining incidence trajectories
or not}
}
\description{
Combine all death trajectories in simulated object
}
