% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_add_trajectory_incidence}
\alias{spim_simulate_add_trajectory_incidence}
\title{Calculate incidence trajectories for a simulated object}
\usage{
spim_simulate_add_trajectory_incidence(obj, states, suffix = "_inc")
}
\arguments{
\item{obj}{A simulated object}

\item{states}{Character vector of states to calculate incidence for}

\item{suffix}{Character giving suffix to add to names of new states, default
"_inc"}
}
\description{
Calculate incidence trajectories for a simulated object
}
