% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_simulate_args}
\alias{spim_simulate_args}
\title{Create simulation parameters}
\usage{
spim_simulate_args(grid, vars, base, ignore, regions, multistrain)
}
\arguments{
\item{grid}{A scenario grid; this is a data.frame with names being
the type of thing being changed and the values representing
quantities in vars}

\item{vars}{A list of parameters organised by scenario}

\item{base}{A list of base parameters}

\item{ignore}{Character vector of additional values in \code{grid} that
are not directly model parameters}

\item{regions}{Character vector of regions}

\item{multistrain}{Logical, indicating if the simulation will use
multiple strains}
}
\value{
A list of parameters for the model
}
\description{
Create a list of simulation parameters
}
