% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_combine_trajectories}
\alias{spim_simulate_combine_trajectories}
\title{Combine regional simulated trajectories}
\usage{
spim_simulate_combine_trajectories(res, name, regions = NULL, rm.rtUK = FALSE)
}
\arguments{
\item{res}{A list of simulated objects}

\item{name}{The name of the aggregated region (e.g. 'england)}

\item{regions}{Character vector of regions to use}

\item{rm.rtUK}{Logical, indicating if the aggregated UK trajectory should be
removed}
}
\description{
Aggregate simulation output by region to produce national estimates
}
