% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_simulate_complete_doses}
\alias{spim_simulate_complete_doses}
\title{Calculate final doses date}
\usage{
spim_simulate_complete_doses(summary)
}
\arguments{
\item{summary}{Simulation summary object}
}
\description{
Calculate when all second doses given out
}
