% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_create_summary}
\alias{spim_simulate_create_summary}
\title{Summarise simulations over particles}
\usage{
spim_simulate_create_summary(x, at = c(1, 20, 39)/40)
}
\arguments{
\item{x}{A simulated object}

\item{at}{Numeric vector of quantiles at which to calculate summary,
defaults to 2.5, 50, 97.5}
}
\description{
Summarise simulations over particles
}
