% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_simulate_local}
\alias{spim_simulate_local}
\title{Run simulations}
\usage{
spim_simulate_local(args, combined)
}
\arguments{
\item{args}{Arguments returned by \link{spim_simulate_args}}

\item{combined}{Processed combined output returned by
\link{spim_simulate_prepare}}
}
\description{
Run simulations locally
}
