% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_simulation.R
\name{spim_simulate_parameter_grid}
\alias{spim_simulate_parameter_grid}
\title{Prepare parameter update list}
\usage{
spim_simulate_parameter_grid(control)
}
\arguments{
\item{control}{A \link{spim_simulate_control} object}
}
\description{
Convert an \link{spim_simulate_control} object into a list of
parmeters for simulation; these will then need to be swapped into
the model too.
}
