% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_simulate_prepare}
\alias{spim_simulate_prepare}
\title{Prepare for simulation}
\usage{
spim_simulate_prepare(
  combined,
  n_par,
  regions = NULL,
  inflate_strain = FALSE,
  inflate_booster = FALSE
)
}
\arguments{
\item{combined}{A "combined" object}

\item{n_par}{The number of parameters to sample}

\item{regions}{Character vector of regions to use}

\item{inflate_strain}{Logical, inidicating if an empty second
strain should be added}

\item{inflate_booster}{Logical, inidicating if an empty booster
dose should be added}
}
\description{
Prepare combined output for simulations. This carries out
maintenance for the state - upgrading state if sircovid has
updated (within reason), sampling starting parameters, reducing to
selected regions and (potentially) adding empty strain or booster
compartments.
}
