% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_process_output}
\alias{spim_simulate_process_output}
\title{Process simulation output}
\usage{
spim_simulate_process_output(
  obj,
  combined_region,
  regions,
  incidence_states,
  reset_states = FALSE,
  rm.rtUK = FALSE
)
}
\arguments{
\item{obj}{A simulated object}

\item{combined_region}{The name of the aggregated region (e.g. 'england)}

\item{regions}{Character vector of regions to use}

\item{incidence_states}{Character vector of states to calculate incidence}

\item{reset_states}{Logical, indicating if incidence states should be reset
to zero at the start of the simulation}

\item{rm.rtUK}{Logical, indicating if the aggregated UK trajectory should be
removed}
}
\description{
Process simulation output
}
