% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_remove_dates_to}
\alias{spim_simulate_remove_dates_to}
\title{Remove simulations up to given date}
\usage{
spim_simulate_remove_dates_to(obj, date)
}
\arguments{
\item{obj}{A simulated object}

\item{date}{If not \code{NULL} then the date to remove all results up
to}
}
\description{
Removes all results from a simulation object up to and including a
given date
}
