% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_reset_cumulative_states}
\alias{spim_simulate_reset_cumulative_states}
\title{Reset cumulative states to zero at start of simulation}
\usage{
spim_simulate_reset_cumulative_states(res, state_names)
}
\arguments{
\item{state_names}{Character vector of states to reset}

\item{obj}{A simulated object}
}
\description{
Reset cumulative states to zero at start of simulation
}
