% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_simulate_rrq}
\alias{spim_simulate_rrq}
\title{Run simulations with rrq}
\usage{
spim_simulate_rrq(args, combined, rrq)
}
\arguments{
\item{args}{Arguments returned by \link{spim_simulate_args}}

\item{combined}{Processed combined output returned by
\link{spim_simulate_prepare}}

\item{rrq}{rrq object}
}
\description{
Run simulations with rrq workers
}
