% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_simplify_rt}
\alias{spim_simulate_simplify_rt}
\title{Simplify simulation object by moving Rt in with trajectories}
\usage{
spim_simulate_simplify_rt(x)
}
\arguments{
\item{x}{A simulated object}
}
\description{
Simplify simulation object by moving Rt in with trajectories
}
