% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_tidy_states}
\alias{spim_simulate_tidy_states}
\title{Create tidy (long) dataframe of simulated results}
\usage{
spim_simulate_tidy_states(res, run_grid)
}
\arguments{
\item{res}{A list of simulated objects}

\item{run_grid}{data.frame with same number of rows as \code{res} giving metadata
to attach to results}
}
\description{
Create tidy (long) dataframe of simulated results
}
