% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_simulation_predictors}
\alias{spim_simulation_predictors}
\title{Return predictive simulation variables}
\usage{
spim_simulation_predictors(summary)
}
\arguments{
\item{summary}{A tidy summary object such as that returned by
\code{create_summary}}
}
\description{
Names of 'predictive' variables from a tidy simulation summary object, i.e.
those variables which: i) are not purely informative;
ii) impact upon predictions; iii) are not outcome variables.
}
