% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_spim.R
\name{spim_summary_nowcast}
\alias{spim_summary_nowcast}
\alias{spim_summary_time_series}
\alias{spim_summary_write}
\title{Create summaries}
\usage{
spim_summary_nowcast(dat, sircovid_model = "carehomes")

spim_summary_time_series(dat, time_series_date, sircovid_model = "carehomes")

spim_summary_write(result, path_template, path_save)
}
\arguments{
\item{dat}{A combined data object (loaded by
\link{spim_combined_load}}

\item{sircovid_model}{The name of the sircovid model used.
Default is \code{"carehomes"}}

\item{time_series_date}{The start date for the preceding time series}

\item{result}{Result of \code{spim_summary_nowcast} or
\code{spim_summary_time_series}}

\item{path_template}{Path to the template file}

\item{path_save}{Path to save the result at}
}
\value{
A \link{data.frame} of results
}
\description{
Create summaries
}
