% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_write_npi_key}
\alias{spim_write_npi_key}
\title{Save NPI key}
\usage{
spim_write_npi_key(npi_key, filename)
}
\arguments{
\item{npi_key}{Output from \link{spim_prepapre_npi_key}}

\item{filename}{File to save NPI key to}
}
\description{
Save NPI key with quantiles, mean, and standard deviation
}
