% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spim_write_uptake}
\alias{spim_write_uptake}
\title{Save dose uptake at a given date}
\usage{
spim_write_uptake(report_date, doses, filename)
}
\arguments{
\item{report_date}{Date to calculate uptake}

\item{doses}{Output from \link{spim_calculate_doses}}

\item{filename}{File to save uptake to}
}
\description{
Write first and second dose uptake at a given date
}
