% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{tidy_state_one}
\alias{tidy_state_one}
\title{Create tidy (long) dataframe of simulated results}
\usage{
tidy_state_one(x, common)
}
\arguments{
\item{x}{A simulated object}

\item{common}{A one-row data.frame giving metadata to attach to results}
}
\description{
Create tidy (long) dataframe of simulated results
}
