% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_calculate_doses}
\alias{spim_calculate_doses}
\title{Calculate doses given out from simulation}
\usage{
spim_calculate_doses(summary, population_england, Scenario)
}
\arguments{
\item{summary}{Output from \link{spim_simulate_tidy_states}}

\item{population}{Vector of England population by 19 age groups
and rows are age groups}
}
\description{
Calculate doses given out from simulation
}
