% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\name{spim_combined_load}
\alias{spim_combined_load}
\title{Load combined fits}
\usage{
spim_combined_load(path, regions = "all", get_severity = FALSE)
}
\arguments{
\item{path}{Directory name. Within here we expect to see
\verb{<region>/fits.rds} for each region in
\code{sircovid::regions("all")}}

\item{regions}{Region type passed to \link[sircovid:regions]{sircovid::regions} (default
is \code{all}, otherwise try \code{england})}

\item{get_severity}{Logical, indicating whether to extract severity (e.g.
IFH, IHR, HFR, etc.) trajectories. Must default to \code{FALSE}, as these
are routine (e.g. MTPs) \code{sircovid} outputs.}
}
\value{
A combined fit object
}
\description{
Load combined fits from a directory of regional fits
}
