% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{spim_control}
\alias{spim_control}
\title{High-level control}
\usage{
spim_control(short_run, n_chains, date_restart = NULL, workers = TRUE)
}
\arguments{
\item{short_run}{Logical, indicating if this is a debug run. This
will use freer particles, mcmc steps and a small sample}

\item{n_chains}{The number of chains to run}

\item{date_restart}{Optionally, dates save restart data in the
pmcmc (see \link[mcstate:pmcmc]{mcstate::pmcmc}}

\item{workers}{Logical, indicating if we should enable workers. If
\code{TRUE}, then a number of workers between 1 and 4 will be used
depending on \code{n_chains} and the detected number of cores.}
}
\value{
A list of options
}
\description{
High-level control.
}
\details{
TODO: document parallel detection
}
