% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{spim_data}
\alias{spim_data}
\title{Organise input data}
\usage{
spim_data(
  date,
  region,
  model_type,
  rtm,
  serology,
  trim_deaths,
  full_data = FALSE,
  fit_to_variants = FALSE
)
}
\arguments{
\item{date}{The date of the data. This will be the final date}

\item{region}{The region that we are fitting to. This will either
be one of the 10 NHS regions (one of
\code{sircovid::regions("all")}), or an argument to that function
(e.g., \code{all})}

\item{model_type}{Either "BB" (beta-binomial) or "NB" (negative
binomial). We use and exclude different data types depending on
this argument.}

\item{rtm}{The "RTM collate" data feed. This contains information
about deaths, positive test results, react serology, etc. TODO:
DESCRIBE FORMAT; TODO: DESCRIBE ORIGIN}

\item{serology}{The additional serology data. TODO: DESCRBE FORMAT}

\item{trim_deaths}{The number of days of deaths to trim to avoid
back-fill issues. We typically use a value of 4 days.}

\item{full_data}{Not sure yet, we'll find out}

\item{fit_to_variants}{Logical, whether to fit to variants data or not}
}
\value{
A \code{\link[=data.frame]{data.frame()}} TODO: describe columns
}
\description{
Organise data inputs that the model will be fit to. Note that
there are other data inputs (notably severity and vaccination
data) that are not fit to but are treated as model parameters.
}
