% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{spim_data_admissions}
\alias{spim_data_admissions}
\title{Load admissions data}
\usage{
spim_data_admissions(new_data, old_data, region)
}
\arguments{
\item{new_data}{A data.frame from the admissions by age sitrep using (finer)
age-bands available from October 13, 2020 onward}

\item{old_data}{A data.frame from the admissions by age sitrep using (broad)
age-bands available from the start of the epidemic onward}

\item{region}{Name of the region}
}
\description{
Process SitRep admissions data
}
