% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_process.R
\name{spim_fit_process}
\alias{spim_fit_process}
\title{Process a fit}
\usage{
spim_fit_process(
  samples,
  control,
  data_admissions,
  rtm,
  parameters,
  data,
  data_full,
  data_vaccination
)
}
\arguments{
\item{samples}{The pmcmc_samples object from \link[mcstate:pmcmc]{mcstate::pmcmc}}

\item{control}{The forecast control from \link{spim_control}}

\item{data_admissions}{The admissions data set from
\link{spim_data_admissions}}

\item{rtm}{The rtm data set}

\item{parameters}{The parameter specification
(\link{spim_pars_pmcmc_load})}

\item{data}{The data set as passed to
\link{spim_particle_filter}}

\item{data_full}{Full data set, before any right-censoring}

\item{data_vaccination}{The vaccination data set as passed to
\link{spim_pars}}
}
\description{
Process fit data
}
