% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_process.R
\name{spim_fit_process}
\alias{spim_fit_process}
\title{Process a fit}
\usage{
spim_fit_process(samples, parameters, data, control)
}
\arguments{
\item{samples}{The pmcmc_samples object from \link[mcstate:pmcmc]{mcstate::pmcmc}}

\item{parameters}{The parameter specification
(\link{spim_pars_pmcmc_load})}

\item{data}{Data sets used in fitting, via
\link{spim_fit_process_data}}

\item{control}{The forecast control from \link{spim_control}}
}
\description{
Process fit data
}
