% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_process.R
\name{spim_fit_process}
\alias{spim_fit_process}
\title{Process a fit for \code{lancelot} model}
\usage{
spim_fit_process(samples, parameters, data, control)
}
\arguments{
\item{samples}{The pmcmc_samples object from \link[mcstate:pmcmc]{mcstate::pmcmc}}

\item{parameters}{The parameter specification
(\link{spim_pars_pmcmc_load})}

\item{data}{Data sets used in fitting}

\item{control}{A list of control parameters including
\code{severity} and \code{compiled_compare}, typically the
\code{particle_filter} element of the result of
\code{\link[=spim_control]{spim_control()}}. It will be used here to check for switches of
particular trajectories to be outputed (e.g. \code{severity}).}
}
\description{
Process fit data for \code{lancelot} model
}
