% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_process.R
\name{spim_fit_process_data}
\alias{spim_fit_process_data}
\title{Collect data sets}
\usage{
spim_fit_process_data(admissions, rtm, fitted, full, vaccination)
}
\arguments{
\item{admissions}{The admissions data set from
\link{spim_data_admissions}. Set to NULL if not fitting or plotting
age-specific data}

\item{rtm}{The rtm data set}

\item{fitted}{The data set as passed to
\link{spim_particle_filter}}

\item{full}{Full data set, before any right-censoring}

\item{vaccination}{The vaccination data set as passed to
\link{spim_pars}}
}
\description{
Collect data sets for use with \link{spim_fit_process}
}
