% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{spim_pars_beta}
\alias{spim_pars_beta}
\title{Create vector of beta dates}
\usage{
spim_pars_beta(date, last_beta_days_ago = 21)
}
\arguments{
\item{date}{Today's dates (last date is set to last_beta_days_ago)}

\item{last_beta_days_ago}{Number of days in the past for last beta point}
}
\description{
Create vector of beta dates
}
