% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{spim_particle_filter}
\alias{spim_particle_filter}
\title{Contruct a particle filter}
\usage{
spim_particle_filter(data, pars, control)
}
\arguments{
\item{data}{Data, from \link{spim_data}}

\item{pars}{Parameters, from \link{spim_pars}}

\item{control}{A list of control parmaeters including
\code{n_particles}, \code{n_threads} and \code{compiled_compare}, typically the
\code{particle_filter} element of the result of
\code{\link[=spim_control]{spim_control()}}}
}
\value{
A \link[mcstate:particle_filter]{mcstate::particle_filter} object
}
\description{
Construct a particle filter
}
