% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{spim_particle_filter}
\alias{spim_particle_filter}
\title{Construct a particle filter}
\usage{
spim_particle_filter(
  data,
  pars,
  control,
  model = "carehomes",
  deterministic = FALSE
)
}
\arguments{
\item{data}{Data, from \link{spim_data}}

\item{pars}{Parameters, from \link{spim_pars}}

\item{control}{A list of control parameters including
\code{n_particles}, \code{n_threads} and \code{compiled_compare}, typically the
\code{particle_filter} element of the result of
\code{\link[=spim_control]{spim_control()}}}

\item{model}{The name of the sircovid model used. Default is \code{"carehomes"}}

\item{deterministic}{Logical, indicating if the particle filter to built
is to be run deterministically or stochastically}
}
\value{
A \link[mcstate:particle_filter]{mcstate::particle_filter} object
}
\description{
Construct a particle filter
}
