% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{spim_particle_filter}
\alias{spim_particle_filter}
\title{Construct a particle filter}
\usage{
spim_particle_filter(
  data,
  pars,
  control,
  deterministic = FALSE,
  initial = NULL,
  initial_date = 0,
  rt = FALSE,
  cum_admit = FALSE,
  diagnoses_admitted = FALSE,
  cum_infections_disag = FALSE,
  cum_n_vaccinated = FALSE
)
}
\arguments{
\item{data}{Data, from \link{spim_data}}

\item{pars}{Parameters, from \link{spim_fit_pars_load}}

\item{control}{A list of control parameters including
\code{n_particles}, \code{n_threads} and \code{compiled_compare}, typically the
\code{particle_filter} element of the result of
\code{\link[=spim_control]{spim_control()}}}

\item{deterministic}{Logical, indicating if the particle filter to built
is to be run deterministically or stochastically}

\item{initial}{Optionally a matrix of model state or an initial
value function.}

\item{initial_date}{The initial date to start from.  Typically
this should be 0 for a parent fit, and be the restart date for a
restart date.}

\item{rt}{logical parameter output trajectories required for
calculating Rt (default = FALSE)}

\item{cum_admit}{logical parameter whether to output cumulative
admissions by age (default = FALSE)}

\item{diagnoses_admitted}{logical parameter whether to output
cumulative combined confirmed admissions and inpatient
diagnoses by age and vaccine class (default = FALSE)}

\item{cum_infections_disag}{logical parameter whether to output
cumulative infections by age and vaccine class (default = FALSE)}

\item{cum_n_vaccinated}{logical parameter whether to output
cumulative number vaccinated by age and vaccine class
(default = FALSE)}
}
\value{
A \link[mcstate:particle_filter]{mcstate::particle_filter} object
}
\description{
Construct a particle filter
}
