% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_plot_check_rt}
\alias{spim_plot_check_rt}
\title{Check Rt from simulation}
\usage{
spim_plot_check_rt(summary_state, combined_state)
}
\arguments{
\item{summary_state}{State from simulation summary object}

\item{combined_state}{State from combined object}
}
\description{
Diagnostic plots for checking Rt from simulation
}
