% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined_forest.R
\name{spim_plot_forest}
\alias{spim_plot_forest}
\title{Forest plot}
\usage{
spim_plot_forest(dat, plot_betas)
}
\arguments{
\item{dat}{Combined data from \link{spim_combined_load}}

\item{plot_betas}{Logical, indicating if we should plot betas
(rather than all other parameters)}
}
\value{
Nothing, called for side effects
}
\description{
Forest plot
}
