% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_log_traj_by_age}
\alias{spim_plot_log_traj_by_age}
\title{Plot log trajectories by age}
\usage{
spim_plot_log_traj_by_age(dat, yield)
}
\arguments{
\item{dat}{Combined data set}

\item{yield}{Which kind of age-specific trajectories to add, can be
"admissions" or "deaths"}
}
\description{
Plot log trajectories by age
}
