% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_pillar2_positivity}
\alias{spim_plot_pillar2_positivity}
\title{Plot Pillar 2 positivity}
\usage{
spim_plot_pillar2_positivity(
  dat,
  regions,
  date_min,
  ymax,
  age_band = NULL,
  add_betas = FALSE
)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{date_min}{Starting date for plot}

\item{ymax}{Maximum percentage on y-axis}

\item{add_betas}{Logical, indicating if we should add betas}
}
\description{
Plot Pillar 2 positivity
}
