% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_react}
\alias{spim_plot_react}
\title{Plot REACT}
\usage{
spim_plot_react(
  dat,
  regions,
  date_min,
  ymax,
  age_band = NULL,
  add_betas = FALSE
)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{date_min}{Starting date for plot}

\item{ymax}{Maximum percentage on y-axis}

\item{age_band}{String, indicating which age band to plot}

\item{add_betas}{Logical, indicating if we should add betas}
}
\description{
Plot REACT
}
