% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_Rt}
\alias{spim_plot_Rt}
\title{Plot Rt}
\usage{
spim_plot_Rt(
  dat,
  regions,
  rt_type,
  forecast_until = NULL,
  variant = NULL,
  add_betas = FALSE
)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{rt_type}{One of the valid Rt types (e.g., eff_Rt_all)}

\item{forecast_until}{Optional date to forecast till}
}
\description{
Plot Rt
}
