% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multivariant.R
\name{spim_plot_seeding_date}
\alias{spim_plot_seeding_date}
\title{Create plot for estimated seeding date}
\usage{
spim_plot_seeding_date(dat)
}
\arguments{
\item{dat}{Main fitting outputs object}
}
\value{
A ggplot2 object for plot of estimated seeding date
}
\description{
Create plot for estimated seeding date
}
