% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{spim_plot_trajectories}
\alias{spim_plot_trajectories}
\title{Plot trajectories}
\usage{
spim_plot_trajectories(
  dat,
  regions,
  what,
  date_min = NULL,
  with_forecast = TRUE,
  add_betas = FALSE,
  sircovid_model = "carehomes"
)
}
\arguments{
\item{dat}{Combined data set}

\item{regions}{Vector of regions to plot}

\item{what}{Vector of traces to add}

\item{date_min}{Starting date for plot. If not given, defaults to
the beginning of the data.}

\item{with_forecast}{Logical, indicating if we should add the forecast}

\item{add_betas}{Logical, indicating if we should add betas}

\item{sircovid_model}{The name of the sircovid model used.
Default is \code{"carehomes"}}
}
\description{
Plot trajectories
}
