% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multivariant.R
\name{spim_plot_voc_proportion}
\alias{spim_plot_voc_proportion}
\title{Create plot for VOC proportion}
\usage{
spim_plot_voc_proportion(dat, date_restart, region)
}
\arguments{
\item{dat}{Main fitting outputs object}

\item{date_restart}{Date for multivariant model restart}

\item{region}{A string for the name of the region to plot}
}
\value{
A ggplot2 object for fit to VOC proportion data
}
\description{
Create plot for VOC proportion
}
