% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{spim_rrq_controller}
\alias{spim_rrq_controller}
\title{Find rrq controller}
\usage{
spim_rrq_controller(root = here::here())
}
\arguments{
\item{root}{Root of the orderly project (used to anchor the rrq
file store).}
}
\value{
Returns an rrq controller object if found, otherwise errors
}
\description{
Finds an rrq controller if available
}
