% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plots.R
\name{spim_scenario_cols}
\alias{spim_scenario_cols}
\title{Return accessible scenario colours}
\usage{
spim_scenario_cols(scens, dark_scens = NULL, darken = 50, palette = "bright")
}
\arguments{
\item{scens}{Unique scenario names}

\item{dark_scens}{Optional unique scenario names that should be same length
as \code{scens} as provided and will be the same colours but darker. Useful
if plotting scenarios and their High R counterparts.}

\item{darken}{If \code{dark_scens} is not \code{NULL} then \code{darken} is subtracted
from the RGB code of \code{palette} to darken the colour palette for
\code{dark_scens}}

\item{palette}{Colour palette, passed to \link[khroma:colour]{khroma::colour}}
}
\description{
Select colours for plotting simulation scenarios
}
