% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tools.R
\name{spim_simulate_add_diagnoses_admitted}
\alias{spim_simulate_add_diagnoses_admitted}
\title{Combine diagnoses and admitted trajectories}
\usage{
spim_simulate_add_diagnoses_admitted(obj, incidence = FALSE)
}
\arguments{
\item{obj}{A simulated object}

\item{incidence}{Logical, whether you are combining incidence trajectories
or not}
}
\description{
Combine diagnoses and admitted trajectories in simulated object
}
