% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{spim_vaccination_data}
\alias{spim_vaccination_data}
\title{Prepare vaccination data}
\usage{
spim_vaccination_data(date, region, uptake, days_to_effect, data)
}
\arguments{
\item{date}{The date of the data. This will be the final date}

\item{region}{The region that we are fitting to. This will either
be one of the 10 NHS regions (one of
\code{sircovid::regions("all")}), or an argument to that function
(e.g., \code{all})}

\item{uptake}{A matrix of proportion uptake by age, with (i, j) th
value being the uptake of dose j for group i. Must have 19 rows.}

\item{days_to_effect}{A vector of the number of days each dose is
assumed to take to reach effect}

\item{data}{Vaccination data (TODO: DESCRIBE CONTENTS)}
}
\value{
A list suitable for passing to \code{spim_pars} as
\code{vaccination}, containing the new vaccination schedule
(important bits are \code{efficacy} and \code{schedule}, but other
elements may be used elsewhere)
}
\description{
Prepare vaccination data
}
