% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{spim_vaccination_data}
\alias{spim_vaccination_data}
\title{Prepare vaccination data}
\usage{
spim_vaccination_data(
  date,
  region,
  uptake,
  end_date,
  mean_days_between_doses,
  efficacy,
  data
)
}
\arguments{
\item{date}{The date of the data. This will be the final date}

\item{region}{The region that we are fitting to. This will either
be one of the 10 NHS regions (one of
\code{sircovid::regions("all")}), or an argument to that function
(e.g., \code{all})}

\item{uptake}{A vector of propotion update by age. Must be length 19}

\item{end_date}{The final date of the simulation; we will expand
out the vaccination schedule to this date (if the simulation
continues beyond that date, it's fine)}

\item{mean_days_between_doses}{The number of days between doses}

\item{efficacy}{Vaccine efficacy data. This will be a list with
elements \code{rel_susceptibility}, \code{rel_p_sympt},
\code{rel_p_hosp_if_sympt}, \code{rel_p_death}, \code{rel_infectivity} (relative effect
of susceptibility, symptomatic disease, hospitalisation conditional
on symtomatic, death conditional on hospitalisation, and infectivity).
Each of the list elements must
be a \verb{19 x n} matrix (where \code{n} is the number of vaccination
classes and is shared across all elements). Every number in the
matrix must be a proportion and decrease across columns.}

\item{data}{Vaccination data (TODO: DESCRIBE CONTENTS)}
}
\value{
A list suitable for passing to \code{spim_pars} as
\code{vaccination}, containing the new vaccination schedule
(important bits are \code{efficacy} and \code{schedule}, but other
elements may be used elsewhere)
}
\description{
Prepare vaccination data
}
